Update `PATH` to include correct bundler version
After upgrading to El Capitan bundler could not be found within `$PATH` which
caused gem installation issues. To fix this, the `$PATH` has been updated to
include the correct path for bundler. Along with this change, I've bumped the
version of bundler used because 1.5.3 is outdated by a couple of years now.

Fixes #794.